Make message type 64 bit to avoid padding

+++ src/net/messages.h
@@ -3,7 +3,7 @@
 
 #include "utl/fundamental_types.h"
 
-enum class message_type : u8
+enum class message_type : u64
 {
     control = 1,
     status = 2
